Adds increased detail of source reconstruction
Adds increased detail about the purpose of the repository, and what/how/why the
Redo Backup and Recovery source code was painstakingly reconstructed.

Further clarifies the differences between the Rescuezilla and "Redo Rescue"
project that was started by the original "Redo Backup and Recovery" author who
resurfaced in June 2020 after a 7.5 year absence.